Changed method of distinguishing table data from section headings
Table data is now determined by whether we've seen a table header
and whether there is an 'Amount' column has any data.

This is pretty crude and fails for the aggregate data pages.
The aggregate data pages are less structured and less interesting than the transaction pages, so this is a suitable compromise for now.